requirements.txt
pandas~=1.5.3
numpy~=1.24.2
streamlit~=1.19.0
plotly~=5.13.1
scipy~=1.10.1
wrds~=3.1.5
PyYAML~=6.0